Add a type expander
This allows to define partial type trees, i.e. type trees where some
items have been defined as primitive types (and that should be defined
as constants).

For example given an object like:
```
{
  "string 1" => "constant 1",
  "string 2" => "constant 2",
  "number 1" => 1234,
  "inner" => {
    "hello" => "world",
    "random" => Randrizer::Types::Int[]
  }
}
```

It will generate a hash in output that has the same structure but where
`hash["inner"]["random"]` has a random integer value.

This is very useful to quickly compose type trees when the structure of
the output is not composed programmatically.